split basic and tool files in test
