+++ libs/agno/tests/integration/models/litellm_openai/test_basic.py
@@ -151,144 +151,3 @@ def test_history():
     assert len(agent.run_response.messages) == 6
     agent.run("Hello 4")
     assert len(agent.run_response.messages) == 8
-
-
-def test_tool_use():
-    """Test tool use functionality with LiteLLM Proxy"""
-    agent = Agent(
-        model=LiteLLMOpenAI(id="gpt-4o"),
-        markdown=True,
-        tools=[DuckDuckGoTools()],
-        telemetry=False,
-        monitoring=False,
-    )
-
-    # Get the response with a query that should trigger tool use
-    response: RunResponse = agent.run("What's the latest news about SpaceX?")
-
-    assert response.content is not None
-    # system, user, assistant (and possibly tool messages)
-    assert len(response.messages) >= 3
-
-    # Check if tool was used
-    tool_messages = [m for m in response.messages if m.role == "tool"]
-    assert len(tool_messages) > 0, "Tool should have been used"
-
-    _assert_metrics(response)
-
-
-@pytest.mark.asyncio
-async def test_async_tool_use():
-    """Test async tool use functionality with LiteLLM"""
-    agent = Agent(
-        model=LiteLLMOpenAI(id="gpt-4o"), markdown=True, tools=[DuckDuckGoTools()], telemetry=False, monitoring=False
-    )
-
-    # Get the response with a query that should trigger tool use
-    response = await agent.arun("What's the latest news about SpaceX?")
-
-    assert response.content is not None
-    # system, user, assistant (and possibly tool messages)
-    assert len(response.messages) >= 3
-
-    # Check if tool was used
-    tool_messages = [m for m in response.messages if m.role == "tool"]
-    assert len(tool_messages) > 0, "Tool should have been used"
-
-    _assert_metrics(response)
-
-
-def test_parallel_tool_calls():
-    """Test parallel tool calls functionality with LiteLLM"""
-    agent = Agent(
-        model=LiteLLMOpenAI(id="gpt-4o"),
-        markdown=True,
-        tools=[DuckDuckGoTools()],
-        telemetry=False,
-        monitoring=False,
-    )
-
-    response = agent.run("What are the latest news about both SpaceX and NASA?")
-
-    # Verify tool usage
-    tool_calls = [msg.tool_calls for msg in response.messages if msg.tool_calls]
-    assert len(tool_calls) >= 1  # At least one message has tool calls
-    assert sum(len(calls) for calls in tool_calls) == 2  # Total of 2 tool calls made
-    assert response.content is not None
-    assert "SpaceX" in response.content and "NASA" in response.content
-    _assert_metrics(response)
-
-
-def test_multiple_tool_calls():
-    """Test multiple different tools functionality with LiteLLM"""
-
-    def get_weather():
-        return "It's sunny and 75°F"
-
-    agent = Agent(
-        model=LiteLLMOpenAI(id="gpt-4o"),
-        markdown=True,
-        tools=[DuckDuckGoTools(), get_weather],
-        telemetry=False,
-        monitoring=False,
-    )
-
-    response = agent.run("What's the latest news about SpaceX and what's the weather?")
-
-    # Verify tool usage
-    tool_calls = [msg.tool_calls for msg in response.messages if msg.tool_calls]
-    assert len(tool_calls) >= 1  # At least one message has tool calls
-    assert sum(len(calls) for calls in tool_calls) == 2  # Total of 2 tool calls made
-    assert response.content is not None
-    assert "SpaceX" in response.content and "75°F" in response.content
-    _assert_metrics(response)
-
-
-def test_tool_call_custom_tool_no_parameters():
-    """Test custom tool without parameters"""
-
-    def get_time():
-        return "It is 12:00 PM UTC"
-
-    agent = Agent(
-        model=LiteLLMOpenAI(id="gpt-4o"),
-        markdown=True,
-        tools=[get_time],
-        telemetry=False,
-        monitoring=False,
-    )
-
-    response = agent.run("What time is it?")
-
-    assert any(msg.tool_calls for msg in response.messages)
-    assert response.content is not None
-    assert "12:00" in response.content
-    _assert_metrics(response)
-
-
-def test_tool_call_custom_tool_untyped_parameters():
-    """Test custom tool with untyped parameters"""
-
-    def echo_message(message):
-        """
-        Echo back the message
-
-        Args:
-            message: The message to echo
-        """
-        return f"Echo: {message}"
-
-    agent = Agent(
-        model=LiteLLMOpenAI(id="gpt-4o"),
-        markdown=True,
-        tools=[echo_message],
-        telemetry=False,
-        monitoring=False,
-    )
-
-    response = agent.run("Can you echo 'Hello World'?")
-
-    assert any(msg.tool_calls for msg in response.messages)
-    assert response.content is not None
-    assert "Echo: Hello World" in response.content
-    _assert_metrics(response)
